requirements.txt
asgiref==3.3.4
Django==3.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-extensions==3.1.5
django-simple-history==3.0.0
pylama==7.6.6
pytest-cov==2.7.1
pytest-django==3.8.0
pytest==5.3.5
isort==4.3.4
mock==2.0.0
eradicate==1.0
structlog==18.2.0
django-dynamic-fixture==2.0.0
django-import-export==2.3.0
mysqlclient==2.1.0
django-extended-choices==1.3.3

# Freeze pyflakes
pyflakes==2.1.1